requirements.txt
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
azure-functions==1.17.0
python-dateutil==2.8.2
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0 